mixins, resources and bukkit stuff

+++ loadercomplex-bukkit/src/main/java/com/enderzombi102/loadercomplex/bukkit/impl/item/BukkitItemStack.java
@@ -47,7 +47,7 @@ public class BukkitItemStack implements ItemStack {
 	}
 
 	@Override
-	public short getDurability() {
+	public int getDurability() {
 		return this.stack.getDurability();
 	}
